Removed chrome_web_usb_service, consolidated into web_usb/actions
