+++ src/store/web_usb/actions.js
@@ -1,27 +1,58 @@
 import _ from 'lodash'
-import WebUsbService from './service'
+import store from '@/store'
+import { REQUEST_DEVICE_FILTERS, READ_MACRO_CONTROL_TRANSFER, WRITE_MACRO_CONTROL_TRANSFER } from './constants'
 const USBDevices = []
 
 // // // //
+// NOTE - these are now CHROME-only WebUSB actions
+// We need to make a separate file for FIREFOX WebUSB actions
 
+// Builds a Vuex-friendly representation of a device
 function buildVuexDevice (usbDevice) {
   return {
     id: usbDevice.id,
     productName: 'AstroKey',
     type: 'chrome_web_usb',
-    opened: usbDevice.opened
+    opened: usbDevice.opened,
+    fetching: false
   }
 }
 
+// Tracks a USBDevice instance in USBDevices
 function trackUsbDevice (device) {
   let trackedDevice = _.find(USBDevices, { id: device.id })
   if (trackedDevice) return
   USBDevices.push(device)
 }
 
+// Pulls a USBDevice instance from USBDevices
 function getUsbDevice (device_id) {
   return _.find(USBDevices, { id: device_id })
 }
+
+function getUniqueId () {
+  return _.uniqueId('CWEB_USB_DEVICE_')
+}
+
+// // // //
+
+// WebUSB Device 'connect' event handler
+navigator.usb.addEventListener('connect', (usbConnectionEvent) => {
+  let device = usbConnectionEvent.device
+  device.id = getUniqueId()
+  trackUsbDevice(device)
+  store.dispatch('web_usb/resetCollcetion')
+})
+
+// WebUSB Device 'disconnect' event handler
+navigator.usb.addEventListener('disconnect', (usbConnectionEvent) => {
+  // USBDevices.push(usbConnectionEvent.device)
+  // TODO - remove from USBDevices
+  // TODO - these should JUST hit the 'device/add' directly
+  // store.commit('web_usb/remove', usbConnectionEvent.device)
+  // USBDevices.push(usbConnectionEvent.device)
+  store.dispatch('web_usb/resetCollcetion')
+})
 
 // // // //
 
@@ -30,10 +61,17 @@ export default {
   // store.dispatch('web_usb/requestDevices')
   // TODO - ensure that this function does not add duplicate devices to store.collection
   requestDevices: ({ dispatch }) => {
-    return WebUsbService.requestDevices()
-    .then((device) => {
-      trackUsbDevice(device)
-      dispatch('resetCollcetion')
+    return new Promise((resolve, reject) => {
+      return navigator.usb.requestDevice({ filters: REQUEST_DEVICE_FILTERS })
+      .then((device) => {
+        trackUsbDevice(device)
+        dispatch('resetCollcetion')
+      })
+      .catch((err) => {
+        console.log('ERR - navigator.usb.requestDevice()')
+        // throw err
+        return reject(err)
+      })
     })
   },
 
@@ -50,10 +88,22 @@ export default {
   // Invoked with:
   // store.dispatch('web_usb/getDevices')
   getDevices: ({ state, commit, dispatch }) => {
-    WebUsbService.getDevices()
-    .then((devices) => {
-      _.each(devices, (d) => { trackUsbDevice(d) })
-      dispatch('resetCollcetion')
+    return new Promise((resolve, reject) => {
+      return navigator.usb.getDevices()
+      .then((deviceArray) => {
+        _.each(deviceArray, (d) => {
+          d.id = getUniqueId()
+          trackUsbDevice(d)
+        })
+
+        dispatch('resetCollcetion')
+        return resolve()
+      })
+      .catch((err) => {
+        console.log('ERR - navigator.usb.getDevices()')
+        // throw err
+        return reject(err)
+      })
     })
   },
 
@@ -63,8 +113,19 @@ export default {
     let usbDevice = getUsbDevice(vuexDevice.id)
     if (!usbDevice) return
 
-    return WebUsbService.openDevice(usbDevice)
-    .then((d) => { dispatch('resetCollcetion') })
+    return new Promise((resolve, reject) => {
+      return usbDevice.open().then(() => {
+        // QUESTION - do we want to manage configuration selection in a separate method?
+        return usbDevice.selectConfiguration(1).then(() => {
+          dispatch('resetCollcetion')
+        })
+      })
+      .catch((err) => {
+        console.log('ERR - USBDevice.open() failure')
+        // throw err
+        return reject(err)
+      })
+    })
   },
 
   // Invoked with:
@@ -73,16 +134,74 @@ export default {
     let usbDevice = getUsbDevice(vuexDevice.id)
     if (!usbDevice) return
 
-    WebUsbService.closeDevice(usbDevice)
-    .then((d) => { dispatch('resetCollcetion') })
+    return new Promise((resolve, reject) => {
+      return usbDevice.close().then(() => {
+        dispatch('resetCollcetion')
+      })
+      .catch((err) => {
+        console.log('ERR - USBDevice.close() failure')
+        // throw err
+        return reject(err)
+      })
+    })
   },
 
   // Invoked with:
   // store.dispatch('web_usb/readMacro', { device: UsbDevice, key: 0x0000 })
-  readMacro: ({ commit }, { device, key }) => WebUsbService.readMacro({ commit }, device, key),
+  readMacro: ({ commit }, { device, key }) => {
+    let usbDevice = getUsbDevice(device.id)
+    if (!usbDevice) return
+
+    // keyIndex in hex: `0x0000`
+    // Returns a Promise to manage asynchonous behavior
+    return new Promise((resolve, reject) => {
+      // Clones the READ_MACRO_CONTROL_TRANSFER request object
+      // And adds custom `value` attribute to handle the index of the key we're reading from
+      let READ_MACRO_OPTIONS = _.clone(READ_MACRO_CONTROL_TRANSFER)
+      READ_MACRO_OPTIONS.value = key
+
+      // NOTE - `device.controlTransferIn` READS DATA FROM DEVICE
+      // TODO - '256' should be '128'
+      // TODO - `256` should be moved into constants.js
+      // QUESTION - what is this `256` again, expected return length?
+      return usbDevice.controlTransferIn(READ_MACRO_OPTIONS, 256)
+      .then((response) => {
+        console.log('readMacro response:')
+        console.log(new Uint8Array(response.data.buffer))
+        return resolve(new Uint8Array(response.data.buffer))
+      })
+      .catch((err) => {
+        console.log('readMacro error:')
+        return reject(err)
+      })
+    })
+  },
 
   // Invoked with:
   // store.dispatch('web_usb/writeMacro', { device: UsbDevice, key: 0x0000, data: [ 1, 2, ... ] })
   // TODO - rename to writeWorkflow
-  writeMacro: ({ commit }, { device, key, data }) => WebUsbService.writeMacro({ commit }, device, key, data)
+  writeMacro: ({ commit }, { device, key, data }) => {
+    let usbDevice = getUsbDevice(device.id)
+    if (!usbDevice) return
+
+    return new Promise((resolve, reject) => {
+      // Clones the READ_MACRO_CONTROL_TRANSFER request object
+      // And adds custom `value` attribute to handle the index of the key we're reading from
+      let WRITE_MACRO_OPTIONS = _.clone(WRITE_MACRO_CONTROL_TRANSFER)
+      WRITE_MACRO_OPTIONS.value = key
+
+      // NOTE - `device.controlTransferOut` WRITES DATA TO DEVICE
+      return usbDevice.controlTransferOut(WRITE_MACRO_OPTIONS, new Uint8Array(data).buffer)
+      .then((response) => {
+        console.log('writeMacro response:')
+        console.log(response)
+        // return resolve(new Uint8Array(response.data.buffer))
+        return resolve(response)
+      })
+      .catch((err) => {
+        console.log('writeMacro error:')
+        return reject(err)
+      })
+    })
+  }
 }
